requirements.txt
openai           
pymupdf            
sqlalchemy         
pydantic           
python-dotenv      
json_repair       
